% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/obcp.R
\name{add_data}
\alias{add_data}
\title{Add new data to the online bayesian changepoint detection model. Used to add new incoming data to the
original model.}
\usage{
add_data(obcp, data)
}
\arguments{
\item{obcp}{The output of the original obcp algorithm}

\item{data}{the data to be added to the model}
}
\description{
Given a new data input, the posterior estimate is calculated, as well as the probability of a changepoint. The
posterior is updated and the max value is recorded in the obcp object.
}
\examples{
model <- obcp(data, 0, 1, 1, 1)
new_data <- c(24.4, 25.5, 26.6)
sapply(new_data, function(x) { add_data(model, x) })
model
}
\keyword{add}
\keyword{data}
\keyword{data,}
\keyword{online,}

