% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/obcp.R
\name{find_changepoints}
\alias{find_changepoints}
\title{Find where the changepoints occured.}
\usage{
find_changepoints(obcp, num)
}
\arguments{
\item{obcp}{the online bayesian changepoint detection model}

\item{num}{the number of changepoints to detect. sorted by p-value}
}
\description{
Detects the location of the changepoints from the data points that the model has seen so far
}
\examples{
model <- obcp(data, 0, 1, 1, 1)
# finds the top three changepoints
find_changepoints(model, 3)
}
\keyword{changepoints}
\keyword{find,}

