% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/obcp.R
\name{obcp}
\alias{obcp}
\title{Performs the algorithm outlined in the paper by Adams and MacKay.}
\usage{
obcp(data, mu0, kappa0, alpha0, beta0)
}
\arguments{
\item{data}{A vector of data to be analyzed. Must be a one-dimensional time series}

\item{mu0}{The prior for mu. The default is set to 0, but it may not be suitable for all data}

\item{kappa0}{The prior for kappa. The default is set to 1.}

\item{alpha0}{The prior for alpha in the inverse-gamma distribution}

\item{beta0}{The prior for beta in the inverse-gamma distribution}
}
\value{
a S3 object containing the r_t, maxes, and other information.
}
\description{
Input must be a vector. The function outputs relevant data from the simulation, including the r_t table
}
\examples{
output <- obcp(data, mu=0, kappa=2, alpha=2, beta=4)
summary(output)
}
\keyword{bayesian,}
\keyword{changepoint,}
\keyword{normal-inverse-gamma}
\keyword{online,}

