% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.obcp_output}
\alias{plot.obcp_output}
\title{Plot the changepoints from the algorithm}
\usage{
\method{plot}{obcp_output}(x)
}
\arguments{
\item{x}{the output from the obcp function}
}
\description{
Plots the changepoints on top of the original data using the ggplot2 library.
}
\examples{
output <- obcp(data)
plot(output)
}
\keyword{plot}

