% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.obcp_output}
\alias{summary.obcp_output}
\title{Get a summary of the online bayesian changepoint output returned from the simulation}
\usage{
\method{summary}{obcp_output}(x)
}
\arguments{
\item{x}{The output of the obcp function. A class of "obcp_output"}
}
\value{
a data frame of relevant info.
}
\description{
Returns a data table of relevatnt information
}
\keyword{obcp}
\keyword{summary,}

